Ticket #5298 - cleanup
